chore: ci-build script
